fix(lint): split service.rs test/trigger family under the 500-line gate
