+++ src/routines/service_slug_tests.rs
@@ -426,107 +426,3 @@ fn svc_update_migrates_workbenches_on_rename() {
     let logs = svc_logs(&store, "rename-id").unwrap();
     assert_eq!(logs, "prior run log");
 }
-
-#[test]
-fn svc_update_sets_ttl_secs() {
-    let _home = TempHome::set();
-    // Covers the `req.ttl_secs` apply branch in `svc_update`.
-    let title = "Svc Update Ttl ZZZ";
-    let store = new_store();
-    let routine = make_routine("ttl-id", title, 1, 1);
-    crate::routine_storage::write_routine(&routine).unwrap();
-    store.lock().unwrap().insert("ttl-id".into(), routine);
-
-    // `with_empty_path` keeps the post-update crontab sync from touching the real
-    // crontab (issue #175): the update succeeds, the sync just warns.
-    // 1800 < the @daily routine's ttl ceiling (min(MAX_TTL_SECS=3600, interval)), so it is a value
-    // that is actually in force rather than one silently clamped down (#468).
-    with_empty_path(|| {
-        let updated = svc_update(
-            &store,
-            "ttl-id",
-            UpdateRoutineRequest {
-                model: None,
-                schedule: None,
-                title: None,
-                agent: None,
-                prompt: None,
-                goal: None,
-                repositories: None,
-                machines: None,
-                enabled: None,
-                ttl_secs: Some(1800),
-                max_runtime_secs: None,
-                tags: None,
-            },
-        )
-        .unwrap();
-        assert_eq!(updated.routine.ttl_secs, Some(1800));
-    });
-}
-
-#[test]
-fn svc_update_sets_max_runtime_secs() {
-    let _home = TempHome::set();
-    // Covers the `req.max_runtime_secs` apply branch in `svc_update`.
-    let title = "Svc Update Max Runtime ZZZ";
-    let store = new_store();
-    let routine = make_routine("max-runtime-id", title, 1, 1);
-    crate::routine_storage::write_routine(&routine).unwrap();
-    store
-        .lock()
-        .unwrap()
-        .insert("max-runtime-id".into(), routine);
-
-    // `with_empty_path` keeps the post-update crontab sync from touching the real
-    // crontab (issue #175): the update succeeds, the sync just warns.
-    with_empty_path(|| {
-        let updated = svc_update(
-            &store,
-            "max-runtime-id",
-            UpdateRoutineRequest {
-                model: None,
-                schedule: None,
-                title: None,
-                agent: None,
-                prompt: None,
-                goal: None,
-                repositories: None,
-                machines: None,
-                enabled: None,
-                ttl_secs: None,
-                max_runtime_secs: Some(1234),
-                tags: None,
-            },
-        )
-        .unwrap();
-        assert_eq!(updated.routine.max_runtime_secs, Some(1234));
-    });
-}
-
-#[test]
-fn svc_update_trims_title_before_persisting() {
-    // Covers the title `.trim()` on the `svc_update` apply path. Renaming with the
-    // same slug but different spacing/case must store the trimmed title.
-    let title = "Svc Update Trim ZZZ";
-    let store = new_store();
-    let routine = make_routine("trim-id", title, 1, 1);
-    crate::routine_storage::write_routine(&routine).unwrap();
-    store.lock().unwrap().insert("trim-id".into(), routine);
-
-    with_empty_path(|| {
-        let updated = svc_update(
-            &store,
-            "trim-id",
-            UpdateRoutineRequest {
-                // Same slug, padded: applies the rename branch without a conflict.
-                title: Some("  Svc Update Trim ZZZ  ".into()),
-                ..empty_update_request()
-            },
-        )
-        .unwrap();
-        assert_eq!(updated.routine.title, title);
-    });
-
-    let _ = crate::routine_storage::remove_routine_dir(&slugify(title));
-}
