+++ src/routines/service_tests.rs
@@ -149,109 +149,6 @@ fn with_working_crontab(body: impl FnOnce()) {
 }
 
 #[test]
-fn svc_list_sorts_by_updated_at() {
-    let _home = TempHome::set();
-    // Covers the `RoutineSort::Updated` arm: sort by `updated_at` ascending.
-    let store = store_with(vec![
-        make_routine("late", "Zeta", 100, 300),
-        make_routine("early", "Alpha", 100, 100),
-        make_routine("mid", "Mid", 100, 200),
-    ]);
-    let query = RoutineListQuery {
-        sort: RoutineSort::Updated,
-        ..Default::default()
-    };
-    let list = svc_list(&store, &query);
-    assert_eq!(list[0].routine.id, "early");
-    assert_eq!(list[1].routine.id, "mid");
-    assert_eq!(list[2].routine.id, "late");
-}
-
-#[test]
-fn svc_list_sorts_by_title_case_insensitively() {
-    let _home = TempHome::set();
-    // Covers the `RoutineSort::Title` arm: sort by lowercased title.
-    let store = store_with(vec![
-        make_routine("banana", "banana", 0, 0),
-        make_routine("apple", "Apple", 0, 0),
-        make_routine("cherry", "CHERRY", 0, 0),
-    ]);
-    let query = RoutineListQuery {
-        sort: RoutineSort::Title,
-        ..Default::default()
-    };
-    let list = svc_list(&store, &query);
-    assert_eq!(list[0].routine.id, "apple");
-    assert_eq!(list[1].routine.id, "banana");
-    assert_eq!(list[2].routine.id, "cherry");
-}
-
-#[test]
-fn svc_list_breaks_ties_on_id_deterministically() {
-    // Routines come off a `HashMap` (unspecified iteration order), so equal
-    // sort keys must be broken on the stable routine id for the listing to be
-    // deterministic. Three routines share a `created_at`; ascending lists them
-    // by id (A→Z) and descending reverses that, never an arbitrary order.
-    let tied = || {
-        store_with(vec![
-            make_routine("charlie", "C", 50, 0),
-            make_routine("alpha", "A", 50, 0),
-            make_routine("bravo", "B", 50, 0),
-        ])
-    };
-
-    let asc = svc_list(&tied(), &RoutineListQuery::default());
-    assert_eq!(
-        asc.iter().map(|resp| &resp.routine.id).collect::<Vec<_>>(),
-        ["alpha", "bravo", "charlie"],
-    );
-
-    let desc = svc_list(
-        &tied(),
-        &RoutineListQuery {
-            order: SortOrder::Desc,
-            ..Default::default()
-        },
-    );
-    assert_eq!(
-        desc.iter().map(|resp| &resp.routine.id).collect::<Vec<_>>(),
-        ["charlie", "bravo", "alpha"],
-    );
-}
-
-#[test]
-fn svc_list_omits_prompt_by_default() {
-    let _home = TempHome::set();
-    // Default query leaves the prompt blank, and `skip_serializing_if` drops the field entirely.
-    let store = store_with(vec![make_routine("a", "Alpha", 0, 0)]);
-    let list = svc_list(&store, &RoutineListQuery::default());
-    assert_eq!(list.len(), 1);
-    assert!(list[0].routine.prompt.is_empty());
-    let json = serde_json::to_value(&list[0]).unwrap();
-    assert!(
-        json.get("prompt").is_none(),
-        "prompt should be absent from the serialized listing, got {json}"
-    );
-}
-
-#[test]
-fn svc_list_includes_prompt_when_requested() {
-    let _home = TempHome::set();
-    let store = store_with(vec![make_routine("a", "Alpha", 0, 0)]);
-    let query = RoutineListQuery {
-        include_prompts: Some(true),
-        ..Default::default()
-    };
-    let list = svc_list(&store, &query);
-    assert_eq!(list[0].routine.prompt, "do the thing");
-    let json = serde_json::to_value(&list[0]).unwrap();
-    assert_eq!(
-        json.get("prompt").and_then(|value| value.as_str()),
-        Some("do the thing")
-    );
-}
-
-#[test]
 fn svc_create_rejects_blank_title() {
     let _home = TempHome::set();
     // Covers the `reject_blank("title", ..)` error arm in `svc_create`: a
